Refactor campaigns overview nodes and improve metrics calculation
This commit does quite a few things to enhance the functionality of the campaigns overview.

Firstly, it removes redundant `isLast` prop from various components including `CampaignsNodeGroup.jsx`, `CampaignsNodeAudience.jsx`, `CampaignsNode.jsx`.

Next, the `CampaignsNodeCampaign.jsx` file was renamed to `CampaignsNodeCampaign.tsx` and was extensively refactored with added `resultRate` and `costPerResult` which replaced `engagementRate` and `costPerEngagement`. The changes make the component more flexible and adaptable to not be constrained to just one type of campaign.

Several necessary adjustments were made in the helper files `resultsHelpers.js` and `campaignsHelpers.ts` to ensure the correct indexes are accessed in the array and to accommodate the refactor in the component.

In addition, the `OverviewNode` was broken down into `OverviewNodeBase`, `OverviewNodeEngageAdSet` and `OverviewNodeTrafficAdSet` to allow for specificity depending on the type of ad set. Changes were done in `overview.ts`

The goal of these changes was to improve the flexibility of the components and enhance the calculation of metrics for different campaigns.