FD-1718 Refactor redundant code and improve type definitions
Removed unnecessary variables and computations that do not impact the core functionality in the file campaignsHelpers.ts. In components, CampaignsNodeGroup.jsx, CampaignsNodeAudience.jsx, and CampaignsNodeCampaign.tsx, the variable "NODE_INDEXES" was not used, hence it was removed. Furthermore, the "Dictionary" type definition was improved by making "T" non-optional in the file common.ts.