FD-1718 "Add checks to ignore deleted campaigns and audiences"
The code has been modified to ignore deleted campaigns and audiences when loading campaign data and getting AdSets. This is to ensure that deleted or removed campaigns and audiences no longer affect the current ones. Another change was made to handle loading of AdSets only for current and non-deleted campaigns, resulting in performance improvement and relevant data retrieval.