FD=1718 Remove unused variables and streamline code
The 'videoIdIndex' in utils.jsx and 'Campaign' in CampaignsLoader.tsx were removed as they were unused. The Sentry module in targetingHelpers.js was reformatted for readability. Code in the same file was streamlined by removing an unnecessary variable and returning the result directly. These changes make the code cleaner and easier to maintain.