FD-1718 Make platform optional and add interest targeting feature flag
The 'platform' field in the 'LookalikeWithPlatform' interface has been updated to be optional, to allow the system to default to both Facebook and Instagram if no platform is set.

In 'CampaignsHeader.tsx', the condition to render the content is updated to show a loading state only when fetching campaign data is in progress.

In 'campaignsHelpers.ts', a parameter 'hasInterestTargetingAccess' is introduced, which is supplied by feature flag 'interest_targeting_enabled'. This controls the display of the 'create audience' node based on if the user has access to interest targeting.

This is also reflected in the 'CampaignsLoader.tsx', where the feature flag is obtained from the current user's artist context. The 'hasInterestTargetingAccess' flag is passed while getting node groups.