FD-1718 Add adAccountId to audience filtering
Added adAccountId parameter to the excludeAudiences function in campaignsHelpers.ts and updated CampaignsLoader.tsx to include an adAccount from facebookIntegration. This was done to improve the accuracy of the audience filtering by taking into account a specific advertising account for each platform. This change will ensure that the audiences associated with the correct ad account are used for campaign targeting, enhancing the relevance and efficiency of ad campaigns.