FD-1718 Add budget status to campaign helper conditions
Introduced a new variable, `hasActiveBudget`, in the getNodeGroups function. This facilitates the checking of budget status before deciding whether a campaign is active. This change is reflected in the edge creation for the types 'group' in the `getEdges` function. Also, added early return in `CampaignsLoader.tsx` for cases when there are no filteredAdSets.